Refactor observeActiveGetDocQueries functionality and restructure file organization
- Moved the `observeActiveGetDocQueries` function to a new file for better organization and clarity.
- Introduced new helper functions `onDocChanged` and `seedCacheFromProvider` to encapsulate specific logic related to document changes and caching.
- Updated import paths in `setupQueryClient` to reflect the new structure.

These changes enhance the modularity and maintainability of the @docukit/docsync2 package, improving the overall developer experience.
